Navigation fixes and modification
