+++ app/src/main/res/layout/fragment_tab.xml
@@ -6,7 +6,7 @@
     app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
     <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/nav_host_fragment_content_main"
+        android:id="@+id/nav_host_tab_fragment"
         android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
@@ -15,13 +15,12 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:navGraph="@navigation/root_nav_graph" />
+        app:navGraph="@navigation/tabs_nav_graph" />
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/bottom_navigation_view"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:visibility="visible"
         app:itemIconTint="@color/nav_item_icon_tint_color"
         app:itemTextColor="@color/nav_item_text_color"
         app:layout_constraintBottom_toBottomOf="parent"
