+++ app/src/main/res/navigation/root_nav_graph.xml
@@ -5,8 +5,6 @@
     android:id="@+id/root_nav_graph"
     app:startDestination="@id/splashFragment">
 
-    <include app:graph="@navigation/main_nav_graph" />
-
     <include app:graph="@navigation/onboarding_nav_graph" />
 
     <action
@@ -57,9 +55,8 @@
         tools:layout="@layout/fragment_pincode" />
 
     <fragment
-        android:id="@+id/homeFragment"
-        android:name="com.unfinished.dsnp_wallet_kotlin.ui.tabs.home.HomeFragment"
-        android:label="fragment_home"
-        tools:layout="@layout/fragment_home" />
+        android:id="@+id/tabFragment"
+        android:name="com.unfinished.dsnp_wallet_kotlin.ui.tabs.TabFragment"
+        android:label="TabFragment" />
 
 </navigation>
