+++ app/src/main/res/navigation/tab_home_nav_graph.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/tab_home_nav_graph"
+    app:startDestination="@id/homeFragment">
+
+    <fragment
+        android:id="@+id/homeFragment"
+        android:name="com.unfinished.dsnp_wallet_kotlin.ui.tabs.home.HomeFragment"
+        android:label="@string/homeFragment"
+        tools:layout="@layout/fragment_home" />
+
+</navigation>
