+++ app/src/main/res/navigation/tab_key_nav_graph.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/tab_key_nav_graph"
+    app:startDestination="@id/keysFragment">
+
+    <fragment
+        android:id="@+id/keysFragment"
+        android:name="com.unfinished.dsnp_wallet_kotlin.ui.tabs.keys.KeysFragment"
+        android:label="@string/keysFragment"
+        tools:layout="@layout/fragment_keys" />
+
+</navigation>
