+++ app/src/main/res/navigation/tab_settings_nav_graph.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/tab_settings_nav_graph"
+    app:startDestination="@id/settingsFragment">
+
+    <include app:graph="@navigation/mnemonic_nav_graph" />
+
+    <fragment
+        android:id="@+id/settingsFragment"
+        android:name="com.unfinished.dsnp_wallet_kotlin.ui.tabs.settings.SettingsFragment"
+        android:label="SettingsFragment"
+        tools:layout="@layout/fragment_settings">
+
+        <action
+            android:id="@+id/action_settingsFragment_to_mnemonic_nav_graph"
+            app:destination="@id/mnemonic_nav_graph"
+            app:enterAnim="@anim/nav_default_enter_anim"
+            app:exitAnim="@anim/nav_default_exit_anim"
+            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
+            app:popExitAnim="@anim/nav_default_pop_exit_anim"/>
+
+    </fragment>
+
+</navigation>
