+++ feature_account/src/main/java/com/unfinished/feature_account/presentation/AccountRouter.kt
@@ -18,7 +18,7 @@ interface AccountRouter : SecureRouter, ReturnableRouter {
 
     fun openConfirmHandleScreen(handle: String?)
     fun openTermsHandleScreen()
-    fun openHomeScreenFromHandle(skip: Boolean, identitySuccess: Boolean)
+    fun openTabScreen(skip: Boolean, identitySuccess: Boolean)
 
     fun openMnemonicScreen(accountName: String?, addAccountPayload: AddAccountPayload)
